chore: return parking location lat long

+++ src/main/java/be/nicholas/api/location/domain/Location.java
@@ -1,4 +1,4 @@
 package be.nicholas.api.location.domain;
 
-public record Location(int latitude, int longitude, String address) {
+public record Location(double latitude, double longitude, String address) {
 }
